+++ nutype_macros/src/any/parse.rs
@@ -27,6 +27,7 @@ pub fn parse_attributes(
         validation,
         new_unchecked,
         const_fn,
+        constructor_visibility,
         default,
         derive_traits,
         derive_unchecked_traits,
@@ -39,6 +40,7 @@ pub fn parse_attributes(
     Ok(Attributes {
         new_unchecked,
         const_fn,
+        constructor_visibility,
         guard,
         default,
         derive_traits,