+++ nutype_macros/src/float/parse.rs
@@ -40,6 +40,7 @@ where
         validation,
         new_unchecked,
         const_fn,
+        constructor_visibility,
         default,
         derive_traits,
         derive_unchecked_traits,
@@ -52,6 +53,7 @@ where
     Ok(Attributes {
         new_unchecked,
         const_fn,
+        constructor_visibility,
         guard,
         default,
         derive_traits,